Move ttlmap to its own module under pkg We want to use this package in the Datadog exporter
