+++ receiver/signalfxreceiver/go.mod
@@ -23,4 +23,6 @@ replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/splun
 
 replace github.com/open-telemetry/opentelemetry-collector-contrib/pkg/batchperresourceattr => ../../pkg/batchperresourceattr
 
+replace github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ttlmap => ../../pkg/ttlmap
+
 replace github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8sclusterreceiver => ../../receiver/k8sclusterreceiver
